refactored build secret replacer into reusable module
